bart-llm/requirements.txt
transformers[torch]
python-dotenv
